laboratory/oos_lab/fta_lab/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
taichi>=1.7.0
pyyaml>=6.0.0
